Moved frontend source data to their own folder. Introduced admin client_src folder.
