+++ client_src/.eslintrc.json
@@ -10,5 +10,5 @@
     "browser": true,
     "es6": true
   },
-  "extends": [ "plugin:react/recommended", "eslint:recommended" ],
+  "extends": [ "plugin:react/recommended", "eslint:recommended" ]
 }
