+++ client_src/frontend/script/reducers/root.js
